pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "seedrank"
version = "0.1.0"
description = "CLI toolkit for programmatic SEO — keyword research, content management, and performance tracking for Claude Code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Seedrank Contributors" },
]
keywords = ["seo", "programmatic-seo", "content", "ai", "claude", "seedrank"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Text Processing :: Markup",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
]

[project.scripts]
seedrank = "seedrank.cli:app"

[project.urls]
Repository = "https://github.com/seedrank/seedrank"
Issues = "https://github.com/seedrank/seedrank/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/seedrank"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/seedrank/cli/init_cmd.py" = ["E501"]  # CLAUDE.md template contains long markdown lines

[tool.pytest.ini_options]
testpaths = ["tests"]
